Adding confirmDialog to confirm deleting of a product and snackBar to show an alert of succesful

+++ src/app/pages/control-panel/control-panel.component.ts
@@ -1,27 +1,27 @@
 import { Component, inject } from '@angular/core';
 import { MatIconModule } from '@angular/material/icon';
-import { MatTooltipModule } from '@angular/material/tooltip';
 import { MatButtonModule } from '@angular/material/button';
 import { MatInputModule } from '@angular/material/input';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatSelectModule } from '@angular/material/select';
 import { MatCardModule } from '@angular/material/card';
+import { FormBuilder, FormGroup, Validators, ReactiveFormsModule } from '@angular/forms';
 
-import { MatSort, MatSortModule } from '@angular/material/sort';
 import { MatTableDataSource, MatTableModule } from '@angular/material/table';
-import { MatPaginator, MatPaginatorModule } from '@angular/material/paginator';
+
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmDialogComponent } from '../../elements/confirm-dialog/confirm-dialog.component';
 
 import { Product } from '../../interfaces/product';
 import { ProductService } from '../../services/product.service';
-import { ProductComponent } from '../../elements/product/product.component';
 import { CommonModule } from '@angular/common';
-import { FormBuilder, FormGroup, Validators, ReactiveFormsModule } from '@angular/forms';
 
+import { MatSnackBar } from '@angular/material/snack-bar';
 
 @Component({
   selector: 'app-control-panel',
   standalone: true,
-  imports: [CommonModule, MatButtonModule, MatIconModule, MatTooltipModule, MatFormFieldModule, MatInputModule, MatSelectModule, MatCardModule, ProductComponent, ReactiveFormsModule, MatTableModule, MatSortModule, MatPaginatorModule],
+  imports: [CommonModule, MatButtonModule, MatIconModule, MatFormFieldModule, MatInputModule, MatSelectModule, MatCardModule, ReactiveFormsModule, MatTableModule],
   templateUrl: './control-panel.component.html',
   styleUrl: './control-panel.component.scss'
 })
@@ -33,7 +33,7 @@ export class ControlPanelComponent {
 
   productForm: FormGroup;
 
-  constructor(private fb: FormBuilder) {
+  constructor(private fb: FormBuilder, private dialog: MatDialog, private _snackBar: MatSnackBar) {
     this.productForm = this.fb.group({
       title: ['', Validators.required],
       price: [0, Validators.min(0)],
@@ -64,14 +64,28 @@ export class ControlPanelComponent {
       this.productList.push(createdProduct);
       this.dataSource.data = this.productList;
       this.productForm.reset();
+      this.showSnackBar('Product added successfully');
     }
     else {
       alert('Please fill out all fields');
     }
   }
   async removeProduct(productId: number) {
-    await this.productService.deleteProduct(productId);
-    this.productList = this.productList.filter(product => product.id !== productId);
-    this.dataSource.data = this.productList;
+    const dialogRef = this.dialog.open(ConfirmDialogComponent);
+
+    dialogRef.afterClosed().subscribe(async result => {
+      if (result === true) {
+        await this.productService.deleteProduct(productId);
+        this.productList = this.productList.filter(product => product.id !== productId);
+        this.dataSource.data = this.productList;
+        this.showSnackBar('Product deleted successfully');
+      }
+    });
   }
+  showSnackBar(mensaje: string) {
+    this._snackBar.open(mensaje, 'Close', {
+      duration: 3000
+    });
+  }
+  
 }
